image: change width/height to size_t and manage storage allocation explicitly

+++ src/mapcraftercore/renderer/image/scaling.h
@@ -20,13 +20,15 @@
 #ifndef IMAGE_SCALING_H_
 #define IMAGE_SCALING_H_
 
+#include <cstddef>
+
 namespace mapcrafter {
 namespace renderer {
 
 class RGBAImage;
 
-void imageResizeSimple(const RGBAImage& image, RGBAImage& dest, int width, int height);
-void imageResizeBilinear(const RGBAImage& image, RGBAImage& dest, int width, int height);
+void imageResizeSimple(const RGBAImage& image, RGBAImage& dest, size_t width, size_t height);
+void imageResizeBilinear(const RGBAImage& image, RGBAImage& dest, size_t width, size_t height);
 void imageResizeHalf(const RGBAImage& image, RGBAImage& dest);
 
 }
